Generalize sort+filter to preprocess+refine
Aside from the name changes, there is one logic change in this commit:
henceforth we don't assume that the filter step keeps candidates in
order, so we move the default candidate to the top of the list after
every re-filter. This allows us to reorder or entirely change the
candidate list after each user input change, if desired. This supports
flx and will also be helpful for find-file.