Improve UX of read-library-name
By actually using our API's text properties correctly instead of doing
a weird hack around them. Specifically, this means that moving exact
matches to the top actually works when leading path components are
included. I needed to rearrange our text property API to make the
highlighting work correctly. The main disadvantage of this approach is
that the disambiguating prefixes are not actually part of the
candidates, so you can't filter them. But I think that's an acceptable
cost for consistency, simplicity, and reliability.