Finally make switch-to-buffer behave reasonably
I'm sorry it took so long. I was really trying to hold out from
porting Ivy's `:preselect' argument. But it's been half a year and I
just cannot get used to the buffer list getting out of order. So I
compromised and implemented a limited form of `:preselect' that I hope
is difficult to use in a way that is bad for predictability.