Inherit from built-in completions faces (#86)
Emacs already has the default faces `completions-annotations` and
`completions-common-part`, the latter of which we were already using
because we queried for completion candidates.

The former face we should inherit from instead of choosing our own color
for `selectrum-completion-annotation`. This increases the chances that
`selectrum-completion-in-region` will look as it should out-of-the-box,
as users' themes are more likely to have customized the built-in faces
than a new face.

We also add the face `selectrum-completion-docsig`, which inherits from
`selectrum-completion-annotation` and is italicized. Because we
distinguish between these two types of information, we provide our own
pair of faces for customization instead of making both use the single
built-in face `completions-annotations`.

Co-authored-by: Radon Rosborough <radon.neon@gmail.com>