Don't show the default candidate in prompt messages.
Using changes suggested in #70, make sure that the default candidate
is not also shown in the prompt message. E.g., "Files (default *.el): ".
Selectrum has its own methods of indicating the default value, making
this method redundant.

This behavior is effectively similar to
`minibuffer-electric-default-mode`, except for that found occurrences
are always removed.