Be more lenient when pressing TAB and nothing matches
Currently, an error occurs when TAB is pressed and no candidate
matches the minibuffer contents.  Moreover, the minibuffer contents
are erased.

This commit makes the behavior more lenient and similar to the default
minibuffer completion -- the bell is rung a “No match” message is
displayed.  Also, ‘completion-fail-discreetly’ is respected so none of
the aforementioned things are done if it is non-nil.