Remove selectrum--when-compile
I guess we don't need it anymore, and Stefan Monnier pointed out that
it does an odd thing, namely using eval with dynamic scoping even
though lexical scoping is enabled in the file. I don't quite
understand the scoping issue but we might as well get rid of a sketchy
seeming function if it's not used anywhere.

We can resurrect it later if we need it again.