Fix nil selectrum--previous-input-string in CRM
When hitting the return key on a candidate while existing candidates have been selected already, I get `(wrong-type-argument char-or-string-p nil)`.

I figured out it's because `selectrum--previous-input-string` is a buffer-local variable, which I think makes it `nil` later on inside `with-temp-buffer` which is used to clean up the candidate list and append on the selected one.

My elisp isn't the best, so while I think this fix works, maybe you can suggest a better one?